refactor: updating content in home page
refactors the content in the home page to better reflect the current activities and interests of the author.

the content was updated to include a more updated bio and additional information about software engineering, open source projects and environmental protection. the body of the page was also updated to include a note about ikigai and a section on the rust programming language.

this change updates the `home.tsx` page by replacing the content with the new content. the diff shows that the age field was updated, the description of the author's profession was changed from "full-time open-source developer" to "software engineer", and the company linked was changed to "defence labs". additionally, the note about networking was removed and replaced with a note about ikigai and a section on rust programming language was added.