refactor home page
refactors the home page by updating two paragraphs and adding a new one. it also updates a heading and adds a link to a dissertation for interested readers.

the first paragraph has been updated to reflect the idea of making ideas useful for someone else, not just for myself.

the second paragraph has been updated to include modern programming languages, green computing, and anything to improve the human-being daily basis activities.

a new paragraph has been added, with a link to a dissertation on how rust and related technology can accomplish the goal of more eco-friendly software.