added new images and updated where.tsx
adds four new images to the src/assets/images folder and updates the `where.tsx` file to add the new images.

the images that were added are `bd-logo.png`, `magalu.jpg`, `anchor.jpg`, and `magalupay.jpg`. in the `where.tsx` file, these images were added as references in the co component, which was used to display information about various companies. the references were also used to update the role, tagline, and what of each company.