added kibana logo
adds a new kibana logo component to the icons.tsx file. this component will provide a logo that can be used to represent the kibana service. the component is written with react and svg, and it has been tested to ensure that it works correctly.

the component is added to the icons.tsx file by adding the following code:

```
export const kibanalogo = () => (
  <svg
    fill="#000000"
    width="256px"
    height="256px"
    viewbox="0 0 32 32"
    xmlns="http://www.w3.org/2000/svg"
  >
    <g id="svgrepo_bgcarrier" stroke-width="0"></g>
    <g id="svgrepo_tracercarrier" stroke-linecap="round" stroke-linejoin="round"></g>
    <g id="svgrepo_iconcarrier">
      {' '}
      <path d="m28.052 31.984h-22.479l13.172-15.812c5.64 3.683 9.307 9.391 9.307 15.812zm28.052 0.052h-23.948v28.735z"></path>{' '}
    </g>
  </svg>
);
```

this component also provides a technology component in the how.tsx file for the kibana logo. this component is added by adding the following code:

```
<technology
  color="#fff"
  icon={<kibanalogo />}
  name="kibana"
  type="frontend framework"
  usecase={'constructing stateful and durable frontends for large and interactive web apps'}
/>
```

this commit also includes some minor changes and bug fixes.