fixed incorrect use case for aws
updates the use case for aws technology to fix a typo in the existing text.

the diff shows a change from `aws lambda as much as possible and mainly sqs/sns to send, store, and receive messages between different applications and services.` to `aws lambdas as much as possible and mainly sqs/sns to send, store and receive messages between different applications and services.`.