update typo in where component
updates a typo in the `where` component. the typo was in the text of the component and was corrected to ensure the proper meaning is conveyed.

the change made was to correct a typo in the `what` property of the `where` component. the typo was corrected from `drived` to `driven`. the change was made to the `where.tsx` file in line 27. the diff of the changes is as follows:

```diff
- rebuilding the it department from scratch drived by creating fast and innovative projects  using the most up-to-date technologies in a serverless environment, releasing new products and integrations, making them also accessible through developer tools.`}
+ rebuilding the it department from scratch driven by creating fast and innovative projects using the most up-to-date technologies in a serverless environment, releasing new products and integrations, making them also accessible through developer tools.`}
```