added support for external funding
adds a new file to the repository, `funding.yml`, which provides users with an easier way to link external funding sources to their projects.

the `funding.yml` file allows users to specify their external funding sources, such as grants, donations, crowdfunding platforms, and more. this will help users better manage their projects and promote their projects to potential funders. the `funding.yml` file will also provide a way for users to indicate to their peers and potential collaborators that their project is funded and that it is a viable project to contribute to.