updating doing component for spotify support
updates the `doing.tsx` component to add support for spotify. this includes updating the logic to set the `spotifyatom` atom with the spotify data from the `doingatom` atom if it exists, and setting a default spotify atom if not. additionally, updating the ui to include the album art, song title, and artist instead of the default hard-coded values.

- updated `doing` component to add support for spotify
- updated the logic to set the `spotifyatom` atom with the spotify data from the `doingatom` atom if it exists, and setting a default spotify atom if not
- updated the ui to include the album art, song title, and artist instead of the default hard-coded values